bitbake.conf: Handle empty BB_CURRENT_MC
Bitbake is about to change the default value of this from "default" to "". The
original reason for this was to make this kind of include file usage easier.
Instead we were going to complicate bitbake code having to map one value into
the other.

Instead, stop using "default" and put a slightly horrible bit of code in bitbake.conf
as an alternative.

This means a "default.conf" in the multiconfig directory will stop working but this
was never something anyone was expected to use.

(From OE-Core rev: 42e314a516b919cf1cda37cd1145fab06eecae2a)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>